% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_metrics.R
\name{classification_metrics}
\alias{classification_metrics}
\title{Compute classification quality metrics}
\usage{
classification_metrics(pred, true)
}
\arguments{
\item{pred}{vector of predicted classes.}

\item{true}{vector of true classes.}
}
\value{
A tibble with global accuracy, and averages (across classes) of
precision, recall and F1-score.
}
\description{
Compute classification quality metrics
}
\examples{
res <- mutate(mtcars, cyl=factor(cyl)) \%>\%
  resample_split(p=0.5) \%>\%
  xgb_fit(resp="cyl", expl=c("mpg", "hp", "qsec"), nrounds=20) \%>\%
  xgb_predict(fns=NULL)
classification_metrics(res$pred, res$cyl)
}
