% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_report.R
\name{classification_report}
\alias{classification_report}
\alias{cr}
\alias{print.cr}
\title{Build a classification report}
\usage{
classification_report(pred, true, exclude = NULL)

cr(pred, true, exclude = NULL)

\method{print}{cr}(x, digits = 3, ...)
}
\arguments{
\item{pred}{vector of predicted classes.}

\item{true}{vector of true classes.}

\item{exclude}{vector of classes to exclude for average metrics computation.
They are then marked by a \code{*}. This is typically used to exclude dominant
classes that skew the average too much.}

\item{x}{output of \code{\link[=classification_report]{classification_report()}}.}

\item{digits}{number of digits to display for each number in the report.}

\item{...}{passed to \code{\link[base:print]{base::print()}}.}
}
\value{
A data.frame with global and per class metrics.
}
\description{
Create a table giving precision, recall and F1-scores.
}
\examples{
# fit and predict a classifier
res <- mutate(mtcars, cyl=factor(cyl)) \%>\%
  resample_cv(k=3) \%>\%
  xgb_fit(resp="cyl", expl=c("mpg", "hp", "qsec"), nrounds=30) \%>\%
  xgb_predict(fns=NULL)
res
classification_report(res$pred, res$cyl)
classification_report(res$pred, res$cyl, exclude=8)
}
