% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\alias{cm}
\title{Build a confusion matrix}
\usage{
confusion_matrix(pred, true)

cm(pred, true)
}
\arguments{
\item{pred}{vector of predicted classes.}

\item{true}{vector of true classes.}
}
\value{
A table object with true values in lines and predicted values in
columns
}
\description{
Build a confusion matrix
}
\examples{
res <- mutate(mtcars, cyl=factor(cyl)) \%>\%
  resample_split(p=0.5) \%>\%
  xgb_fit(resp="cyl", expl=c("mpg", "hp", "qsec"), nrounds=20) \%>\%
  xgb_predict(fns=NULL)
confusion_matrix(res$pred, res$cyl)
}
