% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defactor_refactor.R
\name{defactor}
\alias{defactor}
\title{Convert a factor into integers using 0-based indexing}
\usage{
defactor(x)
}
\arguments{
\item{x}{vector of class factor (or character).}
}
\value{
A vector of integers encoding factors, starting at 0.
}
\description{
Convert a factor (or an object that can be converted as such) into an vector
of integers indexed from 0 (instead of 1 as usual in R). This is what xgboost
expects as input for classification tasks.
}
\examples{
defactor(factor(c("a", "b", "a")))
}
