% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_one_xgb.R
\name{fit_one_xgb}
\alias{fit_one_xgb}
\title{Fit an xgboost model to \emph{one} \code{resamples} object}
\usage{
fit_one_xgb(
  object,
  resp,
  expl,
  params = list(),
  nrounds,
  verbose = 0,
  weight = NULL,
  nthread = 1,
  ...
)
}
\arguments{
\item{object}{of class resamples, created by a \verb{resample_***()} function.}

\item{resp}{name of the response variable.}

\item{expl}{names of the explanatory variables.}

\item{params}{named list of parameters passed to \code{xgboost::xgb.train()}.}

\item{nrounds}{number of boosting rounds (i.e. number of trees).}

\item{verbose}{0 = silent, 1 = display performance, 2 = more verbose.}

\item{weight}{a vector of observation-level weights, one per line of the training set.}

\item{nthread}{number of threads (cores) used to fir each model. This is set
to one by default to avoid conflict with parallelisation per resample (in
\code{\link[=xgb_fit]{xgb_fit()}} which is more efficient). Set this to more than 1 when fitting
only one model.}

\item{...}{other parameters passed to `xgboost::xgb.Train()``}
}
\value{
The input object (a tibble of class \code{resamples}) with an
additional column called \code{model} containing the fitted model object.
}
\description{
This is the "workhorse" function called by the different methods of xgb_fit.
}
\examples{
rs <- resample_identity(mtcars, 1)
m <- fit_one_xgb(object=rs,
  resp="mpg", expl=c("cyl", "hp", "qsec"),
  # pass hyperparameters as a list
  params=list(eta=0.1, max_depth=4),
  nrounds=20
)
m$model[[1]]$params
m <- fit_one_xgb(object=rs,
  resp="mpg", expl=c("cyl", "hp", "qsec"),
  # pass hyperparamters inline, through ...
  eta=0.1, max_depth=4,
  nrounds=20
)
m$model[[1]]$params
}
