% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Compute variable importance}
\usage{
importance(object, cores = 1, ...)
}
\arguments{
\item{object}{an object output by \code{xgb_fit()}, which contains a \code{model} column.}

\item{cores}{integer, number of cores to use for parallel computation.}

\item{...}{passed to \verb{[xgboost::xgb.importance()]}.}
}
\value{
A data.frame with variables and their importance metric.
}
\description{
Compute variable importance
}
\examples{
# fit a model on five bootstraps
m <- resample_boot(mtcars, 5) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
    eta=0.1, max_depth=4, nrounds=20)
# compute variable importance for each model
m <- importance(m)
# look at importance for the first model
m$importance[[1]]
# plot the average importance
plot_importance(m)
}
\seealso{
Other variable importance functions: 
\code{\link{plot_importance}()}
}
\concept{variable importance functions}
