% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority_vote.R
\name{majority_vote}
\alias{majority_vote}
\title{Return the majority vote in a discrete valued vector}
\usage{
majority_vote(x)
}
\arguments{
\item{x}{a vector coercible to a factor.}
}
\value{
The most common value in x, as a factor.
}
\description{
Return the majority vote in a discrete valued vector
}
\examples{
majority_vote(c(1,2,3,3,3))
majority_vote(factor(mtcars$cyl))
}
