% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_grid.R
\name{param_grid}
\alias{param_grid}
\title{Define a parameter grid, to be explored through the fitting of resamples}
\usage{
param_grid(object, ...)
}
\arguments{
\item{object}{of class resamples, created by a \verb{resample_***()} function.}

\item{...}{named vectors of parameters, the combinations of which will
constitute the grid}
}
\value{
An object of class \code{resamples_grid}, with one row per combination of
parameters value x resample (i.e. row of object). This can then be input to
\code{xgb_fit()} for training.
}
\description{
Define a parameter grid, to be explored through the fitting of resamples
}
\examples{
rs <- resample_boot(mtcars, n=2)
param_grid(rs, eta=c(0.1, 0.5), max_depth=c(2, 4))
}
