% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{plot.cm}
\alias{plot.cm}
\title{Plot a confusion matrix}
\usage{
\method{plot}{cm}(x, trans = "log1p", ...)
}
\arguments{
\item{x}{a confusion matrix built by \code{\link[=confusion_matrix]{confusion_matrix()}}.}

\item{trans}{transformation function for the color scale.}

\item{...}{passed to \code{\link[base:plot]{base::plot()}}.}
}
\value{
A ggplot object showing a heatmap with true values in lines and
predicted values in columns.
}
\description{
Plot a confusion matrix
}
\examples{
res <- mutate(mtcars, cyl=factor(cyl)) \%>\%
  resample_split(p=0.5) \%>\%
  xgb_fit(resp="cyl", expl=c("mpg", "hp", "qsec"), nrounds=20) \%>\%
  xgb_predict(fns=NULL)
plot(confusion_matrix(res$pred, res$cyl))
}
