% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{plot_importance}
\alias{plot_importance}
\title{Plot variable importance}
\usage{
plot_importance(object)
}
\arguments{
\item{object}{an object output by \code{importance()}, which contains an \code{importance} column.}
}
\value{
A data.frame with variables and their importance metric.
}
\description{
Plot variable importance
}
\examples{
# fit a model on five bootstraps
m <- resample_boot(mtcars, 5) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
    eta=0.1, max_depth=4, nrounds=20)
# compute variable importance for each model
m <- importance(m)
# plot the average importance
plot_importance(m)
}
\seealso{
Other variable importance functions: 
\code{\link{importance}()}
}
\concept{variable importance functions}
