% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{plot_partials}
\alias{plot_partials}
\title{Plot (and summarise) partial dependence plots}
\usage{
plot_partials(
  object,
  fns = list(location = mean, spread = stats::sd),
  rug = TRUE
)
}
\arguments{
\item{object}{an object output by \verb{[partials()]}, which contains a \code{partial}
column.}

\item{fns}{a list of summary functions; one should be called \code{location} and
be used to compute the central location of the variable (e.g., mean,
median, etc.); another should be called \code{spread} and be used to compute the
spread around that location (e.g., sd, mad, etc.). When \code{fns} is \code{NULL},
the partial dependence line for each model is plotted.}

\item{rug}{boolean; whether to add a rug plot to show at which values of the
explanatory variables the partial dependence is computed. This is most
useful when partial dependence is computed at quantiles of the original
data (\code{quantiles=TRUE} in \verb{[partials()]}).}
}
\value{
A ggplot2 object.
}
\description{
Plot (and summarise) partial dependence plots
}
\examples{
# fit a model on five bootstraps
m <- resample_boot(mtcars, 5) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
    eta=0.1, max_depth=4, nrounds=20)
# plot all five partial dependence lines
partials(m, expl=c("hp", "cyl")) \%>\%
  plot_partials(fns=NULL, rug=FALSE)
# plot mean+/-sd among resamples, at given quantiles of the training data
partials(m, expl=c("hp", "cyl"), quantiles=TRUE) \%>\%
  plot_partials()
# do the same with median+/-mad
partials(m, expl=c("hp", "cyl"), quantiles=TRUE) \%>\%
  plot_partials(fns=list(location=median, spread=mad))
}
\seealso{
Other partial dependence plots functions: 
\code{\link{partials}()}
}
\concept{partial dependence plots functions}
