% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_metrics.R
\name{pred_metrics}
\alias{pred_metrics}
\title{Compute prediction quality metrics}
\usage{
pred_metrics(y_pred, y_true)
}
\arguments{
\item{y_pred}{numeric vector of predictions}

\item{y_true}{numeric vector of actual values}
}
\value{
a tibble with RMSE, MAE, R2, correlation R2, correlation R2 on log(n+1) transformed data
}
\description{
Compute prediction quality metrics
}
\examples{
res <- resample_split(mtcars, p=0.8) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"), nrounds=100) \%>\%
  xgb_predict(fns=c())
pred_metrics(res$pred, res$mpg)
}
