% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defactor_refactor.R
\name{refactor}
\alias{refactor}
\title{Convert a vector of 0-based indexed integers into a factor}
\usage{
refactor(x, levels)
}
\arguments{
\item{x}{vector of integers representing a factor, with 0-based indexing.}

\item{levels}{levels of the factor to recreate, given in order.}
}
\value{
A vector of class factor
}
\description{
Convert a 0-based indexed vector, which is the output of xgboost for
classification, into an R factor.
}
\examples{
xf <- factor(c("a", "b", "a"))
xi <- defactor(xf)
xi
xf2 <- refactor(xi, levels=c("a", "b"))
identical(xf, xf2)
}
