% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{regression_metrics}
\alias{regression_metrics}
\alias{pred_metrics}
\title{Compute regression quality metrics}
\usage{
regression_metrics(pred, true)

pred_metrics(pred, true)
}
\arguments{
\item{pred}{numeric vector of predictions}

\item{true}{numeric vector of actual values}
}
\value{
a tibble with RMSE, MAE, R2, correlation R2, correlation R2 on log(n+1) transformed data
}
\description{
Compute regression quality metrics
}
\section{Deprecated}{

\code{\link[=pred_metrics]{pred_metrics()}} is deprecated; use \code{\link[=regression_metrics]{regression_metrics()}} instead.
}

\examples{
res <- resample_split(mtcars, p=0.8) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"), nrounds=100) \%>\%
  xgb_predict(fns=NULL)
regression_metrics(res$pred, res$mpg)
}
