% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate}
\alias{replicate}
\title{Replicate each row of a resamples object}
\usage{
replicate(object, n, ...)
}
\arguments{
\item{object}{of class resamples, created by a \verb{resample_***()} function.}

\item{n}{number of times to replicate}
}
\description{
This is useful to test the replicability of a modelling pipeline or to prepare for permutations.
}
\examples{
rs <- resample_split(mtcars, p=0.7)
replicate(rs, n=5)
rs <- resample_cv(mtcars, k=3, n=2)
replicate(rs, n=3)
}
