% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_boot.R
\name{resample_boot}
\alias{resample_boot}
\title{Generate data resamples using bootstrap}
\usage{
resample_boot(data, n = 10, ...)
}
\arguments{
\item{data}{data.frame, the data to resample.}

\item{n}{integer, number of bootstraps.}

\item{...}{unquoted names of columns of .data to stratify by. Usually they are discrete variables.}
}
\value{
A tibble with columns
\itemize{
\item train : an object of class modelr::resample. The boostrapped data.
\item val   : an object of class modelr::resample. The out-of-bag data (i.e. the
rows of .data not selected in the current bootstrap).
\item boot  : integer, the bootstrap number.
}
}
\description{
Generate data resamples using bootstrap
}
\examples{
resample_boot(mtcars, n=2)
rs <- resample_boot(mtcars, n=2)
data.frame(rs$train[1])
# = some rows are repeated
data.frame(rs$val[1])
# = all these rows are not above

# stratified bootstrap
rs  <- resample_boot(mtcars, n=10)
rss <- resample_boot(mtcars, n=10, gear)
sapply(rs$train, function(x) {sum(data.frame(x)$gear==4)})
# = variable number of occurrence of gear==4 in the training set
sapply(rss$train, function(x) {sum(data.frame(x)$gear==4)})
# = reliable number of gear==4 in the training set
}
