% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_identity.R
\name{resample_identity}
\alias{resample_identity}
\title{Generate repeated resamples of the same data}
\usage{
resample_identity(data, n = 10)
}
\arguments{
\item{data}{data.frame, the data to resample.}

\item{n}{integer, number of repetitions.}
}
\value{
A tibble with columns
\itemize{
\item train : an object of class modelr::resample. The input data with reshuffled rows.
\item val   : an empty object of class modelr::resample.
\item id    : integer, the repetition number.
}
}
\description{
This is a dummy function, used to check the consistency of further computation.
It just reshuffles the rows of the input data.
}
\examples{
rs <- resample_identity(mtcars, n=3)
rs
data.frame(rs$train[1])
data.frame(rs$train[2])
# = same except for the row order
}
