% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_split.R
\name{resample_split}
\alias{resample_split}
\title{Generate train-val splits of the data}
\usage{
resample_split(.data, ..., p = 0.8, n = 1)
}
\arguments{
\item{.data}{data.frame, the data to split.}

\item{...}{unquoted names of columns of .data to stratify by. Usually they are discrete variables.}

\item{p}{in [0,1], the proportion of observations to use for training.}

\item{n}{integer, number of repetitions of the split.}
}
\value{
A tibble with columns
\itemize{
\item train : an object of class modelr::resample. The training data.
\item val   : an object of class modelr::resample. The validation data (i.e. the
rows of .data not selected in the training data).
\item rep   : integer, the repetition number.
}
}
\description{
Generate train-val splits of the data
}
\examples{
resample_split(mtcars, p=0.7, n=5)

# stratify train-val by gear
rs  <- resample_split(mtcars, p=0.5, n=10)
rss <- resample_split(mtcars, p=0.5, n=10, gear)
sapply(rs$train, function(x) {sum(data.frame(x)$gear==4)})
# = variable number of occurrence of gear==4 in the training set
sapply(rss$train, function(x) {sum(data.frame(x)$gear==4)})
# = reliable number of gear==4 in the training set
}
