% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error.R
\name{se}
\alias{se}
\title{Compute the standard error of the mean, assuming a normal distribution}
\usage{
se(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
The value of the standard error of the mean = sd / sqrt(n)
}
\description{
Compute the standard error of the mean, assuming a normal distribution
}
