% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_cv.R
\name{split_in_folds}
\alias{split_in_folds}
\title{Split n items into k folds}
\usage{
split_in_folds(n, k)
}
\arguments{
\item{n}{integer, the number of items.}

\item{k}{integer, the number of folds.}
}
\value{
A vector of length n, containing a integer in 1:k = the fold identifier
}
\description{
Split n items into k folds
}
\examples{
joml:::split_in_folds(k=3, n=10)
table(joml:::split_in_folds(k=3, n=100))
}
