% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{summarise_partials}
\alias{summarise_partials}
\title{Summarise partial dependence across resamples}
\usage{
summarise_partials(object, fns = list(location = mean, spread = stats::sd))
}
\arguments{
\item{object}{an object output by \verb{[partials()]}, which contains a \code{partial}
column.}

\item{fns}{a list of summary functions; one should be called \code{location} and
be used to compute the central location of the variable (e.g., mean,
median, etc.); another should be called \code{spread} and be used to compute the
spread around that location (e.g., sd, mad, etc.). When \code{fns} is \code{NULL},
the partial dependence is just concatenated across resamples.}
}
\value{
A data.frame with:
\itemize{
\item \code{variable}: the variable whose dependence to is computed;
\item \code{value}: the value of the variable at which the model marginal effects
are computed.
\item \code{yhat} or \code{yhat_loc}+\code{yhat_spr}: the average prediction of the model for
this value. either as is or the summary of its location (\code{loc}) and spread
(\code{spr}) according to the functions in \code{fns}.
}
}
\description{
Summarise partial dependence across resamples
}
\examples{
# fit a model on 5 bootstraps
m <- resample_boot(mtcars, 5) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
    eta=0.1, max_depth=4, nrounds=20)
# assess variable importance
importance(m) \%>\% summarise_importance()

# compute the partial dependence to the two most relevant variables
m <- partials(m, expl=c("hp", "cyl"))
# and plot them for each resample
plot_partials(m, fns=NULL)
# do the same with a finer grid
m <- partials(m, expl=c("hp", "cyl"), grid.resolution=50)
plot_partials(m, fns=NULL)
# or along quantiles
m <- partials(m, expl=c("hp", "cyl"), quantiles=TRUE, probs=0:20/20)
plot_partials(m, fns=NULL)

# compute mean+/-sd among resamples
summarise_partials(m)
plot_partials(m)
# do the same with median+/-mad
summarise_partials(m, fns=list(location=median, spread=mad))
plot_partials(m, fns=list(location=median, spread=mad))
}
\seealso{
Other partial dependence plots functions: 
\code{\link{partials}()},
\code{\link{plot_partials}()}
}
\concept{partial dependence plots functions}
