% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_fit.R
\name{xgb_fit}
\alias{xgb_fit}
\title{A generic function to fit an xgboost model}
\usage{
xgb_fit(object, ...)
}
\arguments{
\item{object}{of class resamples.}

\item{...}{passed to further methods.}
}
\description{
A generic function to fit an xgboost model
}
