% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_fit.R
\name{xgb_fit.resamples}
\alias{xgb_fit.resamples}
\title{Fit an xgboost model for each row of a \code{resamples} object}
\usage{
\method{xgb_fit}{resamples}(object, cores = 1, ...)
}
\arguments{
\item{object}{of class resamples, created by a \verb{resample_***()} function.}

\item{cores}{integer, number of cores to use for parallel computation.}

\item{...}{passed to \code{fit_one_xgb()} and then later to \code{xgboost::xgb.Train()}}
}
\value{
Like \code{fit_one_xgb()} but with several lines.
}
\description{
Fit an xgboost model for each row of a \code{resamples} object
}
\examples{
resample_boot(mtcars, 3) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
    eta=0.1, max_depth=4, nrounds=20)
}
