% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_fit.R
\name{xgb_fit.resamples_grid}
\alias{xgb_fit.resamples_grid}
\title{Fit an xgboost model for each row of a \code{resamples.grid} object}
\usage{
\method{xgb_fit}{resamples_grid}(object, cores = 1, ...)
}
\arguments{
\item{object}{of class resamples, created by the \code{param_grid()} function.}

\item{cores}{integer, number of cores to use for parallel computation.}

\item{...}{passed to \code{fit_one_xgb()} and then later to \code{xgboost::xgb.Train()}}
}
\value{
The input object with an additional column called
\code{model} containing the fitted model object.

Like fit_one_xgb() but with several lines and grouped per parameter
combination defined in the grid.
}
\description{
Fit an xgboost model for each row of a \code{resamples.grid} object
}
\examples{
m <- resample_boot(mtcars, 2) \%>\%
  param_grid(eta=c(0.1, 0.5), max_depth=c(2, 4)) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
          min_child_weight=2, nrounds=20)
m$model[[1]]$params
m$model[[3]]$params
}
