% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_summarise_fit.R
\name{xgb_summarise_fit}
\alias{xgb_summarise_fit}
\title{Summarise the fit of xgboost models over resamples}
\usage{
xgb_summarise_fit(object, fns = list(mean = mean, sd = stats::sd, se = se))
}
\arguments{
\item{object}{an object output by \code{xgb_fit()}, which contains a \code{model} column.}

\item{fns}{a named list of summary functions, to compute for each
error metric in the evaluation log. If NULL or an empty list,
just return the full log.}
}
\value{
A tibble with columns
\itemize{
\item the grouping columns in \code{object}. Ungroup the object before \code{xgb_summarise_fit()}
if this is not the desired behaviour.
\item \code{iter}    : the iteration, i.e. boosting round
\item \verb{***_+++} : where *** is the metric (e.g. val_rmse for the RMSE on the
validation set) and +++ is the summary function (e.g. mean)
If the input was grouped, the grouping variables are preserved. Ungroup the
object beforehand if this is not the desired behaviour.
}
}
\description{
For each model in the \code{resamples} object, extract the evaluation log, created
during training. It gives the error metric as a function of the number of
boosting rounds. Compute summary statistics over the various resamples.
}
\examples{
# compute summary across resamples
resample_boot(mtcars, n=2) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
          eta=0.1, max_depth=2, nrounds=10) \%>\%
  xgb_summarise_fit(fns=list(mean=mean, sd=sd))

# when using a grid of parameters, the corresponding values of the parameters
# are preserved in the output.
x <- resample_boot(mtcars, n=2) \%>\%
  param_grid(eta=c(0.5, 1)) \%>\%
  xgb_fit(resp="mpg", expl=c("cyl", "hp", "qsec"),
          max_depth=2, nrounds=10) \%>\%
  xgb_summarise_fit()
print(x)
\dontrun{
ggplot(x) + geom_path(aes(x=iter, y=val_rmse_mean, colour=factor(eta)))
}
}
